Bookings!
Improved Booking model (status,status id, customer. customer id ManyToOne relations) and added initial functionality to BookingController - code is in a stable and working state.
Updated Status & StatusUtil

Added utilities.Predicates class for future filtering option - on second thought, might use concatinated query strings instead (could minimize results size sent from db), but it's a "good to have" functionality just in case additional filtering will be required.

+++ conf/schema.sql
@@ -32,7 +32,7 @@ CREATE TABLE booking
 	equipment VARCHAR(255), 
 	camera_settings VARCHAR(255),
 	optimal_lighting_spots VARCHAR(255),
-	status_id INT NOT NULL,
+	status_id BIGINT NOT NULL,
 	num_pics INT,
 	num_selected INT,
 	num_processed INT,
